// cgocore-ui/src/app/features/models/bookings/views/basic/BookingBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class BookingBasicView {

id: number;
selected: boolean;
canDelete: boolean;
airWaybillAirWaybillNumber: string;
originIataCode: string;
destinationIataCode: string;
numberOfPieces: number;
grossWeight: number;
grossWeightUnitOfMeasureCode: string;
grossVolume: number;
grossVolumeUnitOfMeasureCode: string;

/* Specific Code Start */
/* Specific Code End */
}