// cgocore-ui/src/app/features/models/bookings/filters/BookingFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class BookingFilter {

airWaybillAirWaybillNumber: string;
originIataCode: string;
destinationIataCode: string;

/* Specific Code Start */
/* Specific Code End */
}